1.08 can show candle kline in detailActivity

+++ app/src/main/java/com/udacity/stockhawk/bean/KLineBean.java
@@ -1,7 +1,5 @@
 package com.udacity.stockhawk.bean;
 
-import java.math.BigDecimal;
-
 /**
  * author：ajiang
  * mail：1025065158@qq.com
@@ -9,9 +7,9 @@ import java.math.BigDecimal;
  */
 public class KLineBean {
     public long date;
-    public BigDecimal open;
-    public BigDecimal close;
-    public BigDecimal high;
-    public BigDecimal low;
+    public float open;
+    public float close;
+    public float high;
+    public float low;
     public Long vol;
 }
